@@ -19,7 +19,7 @@
     }
   ],
   "require": {
-    "php": "^7.0 || ^8.0",
+    "php": "^7.2 || ^8.0",
     "kint-php/kint": "^3.2",
     "guzzlehttp/guzzle": ">=6.2.0",
     "ext-json": "*"
@@ -31,7 +31,7 @@
     "classmap": ["lib/", "lib/error/", "models/", "models/field/"]
   },
   "require-dev": {
-    "phpunit/phpunit": "^4.0",
+    "phpunit/phpunit": "^9.5.8",
     "php-coveralls/php-coveralls": "^2.1"
   }
 }
