@@ -1,7 +1,7 @@
 <?php
-class PodioAppFieldCollectionTest extends PHPUnit_Framework_TestCase {
+class PodioAppFieldCollectionTest extends \PHPUnit\Framework\TestCase {
 
-  public function setup() {
+  public function setUp() {
     $this->collection = new PodioAppFieldCollection(array(
       new PodioAppField(array('field_id' => 1, 'external_id' => 'a', 'type' => 'text')),
       new PodioAppField(array('field_id' => 2, 'external_id' => 'b', 'type' => 'number')),
